+++ contrib-jam/src/main/java/net/mostlyoriginal/api/component/graphics/RendererSingleton.java
@@ -1,0 +1,14 @@
+package net.mostlyoriginal.api.component.graphics;
+
+import com.artemis.Component;
+import com.artemis.annotations.Transient;
+import net.mostlyoriginal.api.Singleton;
+
+/**
+ * @author Daan van Yperen
+ */
+@Singleton
+@Transient
+public class RendererSingleton extends Component {
+    public boolean sortedDirty = true;
+}